updated delete n put routes

+++ app.js
@@ -33,7 +33,9 @@ app.post("/register", (req, res) => {
 app.post("/login", (req, res) => {
     let email = req.body.email;
     let password = req.body.passcode;
-
+    if(!email | !password){
+        res.send("Field Empty");
+    }
     let sql = `SELECT * FROM users where userID = ?;`;
     pool1.query(sql, email, function(err, rows){
         let userID = rows[0].id;
@@ -55,6 +57,9 @@ app.post("/login", (req, res) => {
 app.post("/blog/:userID", function(req, res){
     let title = req.body.title;
     let body = req.body.content;
+    if(!title | !body){
+        res.send("Field Empty");
+    }
     let userID = req.params.userID;
     let sql = `INSERT INTO blogs (title, body, userid) VALUES (?, ?, ?);`;
     pool.query(sql, [title, body, userID], function(err, rows) {
@@ -64,30 +69,54 @@ app.post("/blog/:userID", function(req, res){
 })
 
 // update blog
-app.put("/blog/:blogID", (req, res) => {
+app.put("/blog/:userID/:blogID", (req, res) => {
     let title = req.body.title;
     let body = req.body.content;
+    let ID = req.params.blogID;
+    let userID = req.params.userID;
     if(!title | !body){
         res.send("Field Empty");
     }
-    let ID = req.params.blogID;
-    let sql = `UPDATE blogs SET ? WHERE id = ID`;
-    pool.query(sql,{title: title, body: body}, function(err, rows) {
+    let sql = `SELECT userid from blogs where id = ID`;
+    pool.query(sql, function(err, rows){
         if(err) throw err;
-        res.send(rows);
+        if(userID == rows[0].userid){
+            sql = `UPDATE blogs SET ? WHERE id = ID and userid = userID`;
+            pool.query(sql,{title: title, body: body}, function(err, rows) {
+                if(err) throw err;
+                res.send(rows);
+            })
+        }else{
+            res.send("Not your blog");
+        }
     })
 })
 
 // delete blog
-app.delete("/blog/:blogID", (req, res) => {
+app.delete("/blog/:userID/:blogID", (req, res) => {
     let ID = req.params.blogID;
-    var sql = "DELETE FROM blogs WHERE id = ID";
-    pool.query(sql, function(err, rows) {
+    let userID = req.params.userID;
+    let sql = `SELECT userid from blogs where id = ID`;
+
+    pool.query(sql, function(err, rows){
         if(err) throw err;
-        res.send(rows);
+        if(userID == rows[0].userid){
+            sql = "DELETE FROM blogs WHERE id = ID";
+            pool.query(sql, function(err, rows) {
+                if(err) throw err;
+                res.send("Successfully deleted");
+            })
+        }else{
+            res.send("Not your blog");
+        }
     })
 })
 
 app.listen(3000, function(){
     console.log("Am ready n running");
 })
+
+
+// create table vs code se hoga
+// authentication achha krma
+// update ko update kro
